backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alembic==1.13.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
async-timeout==4.0.3
asyncpg==0.29.0
attrs==25.4.0
bcrypt==4.0.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cohere==5.20.6
cryptography==46.0.5
dataclasses-json==0.6.7
deprecation==2.1.0
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.1.0
exceptiongroup==1.3.1
fastapi==0.109.2
fastavro==1.12.1
filelock==3.24.3
frozenlist==1.8.0
fsspec==2026.2.0
google-auth==2.48.0
gotrue==2.12.4
greenlet==3.3.1
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.26.0
huggingface_hub==1.4.1
hyperframe==6.1.0
idna==3.11
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.1.9
langchain-community==0.0.38
langchain-core==0.1.53
langchain-openai==0.1.5
langchain-postgres==0.0.9
langsmith==0.1.147
lxml==6.0.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
multidict==6.7.1
mypy_extensions==1.1.0
numpy==1.26.4
openai==1.12.0
orjson==3.11.7
packaging==23.2
passlib==1.7.4
pgvector==0.2.5
postgrest==1.1.1
propcache==0.4.1
psycopg==3.3.2
psycopg-pool==3.3.0
psycopg2-binary==2.9.11
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.1.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.11.0
pypdf==6.7.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.3
realtime==2.5.3
redis==7.2.0
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.2
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.27
starlette==0.36.3
storage3==0.12.1
StrEnum==0.4.15
supabase==2.16.0
supafunc==0.10.2
tenacity==8.5.0
tiktoken==0.12.0
tokenizers==0.22.2
tqdm==4.67.3
typer==0.24.0
typer-slim==0.24.0
types-requests==2.32.4.20260107
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.27.1
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
yarl==1.22.0
langfuse>=2.0,<3.0
pytest==8.3.5
pytest-asyncio==0.25.3
aiosqlite==0.21.0
